APIConnector 공통코드 부모로 추출
